Fix setup wizard logic if new onboarding is presented
